+++ tests/test_runtime.py
@@ -16,7 +16,7 @@
 # OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
 import pytest
-from statechart import InitialState, Metadata, State, Statechart, Transition
+from statechart import InitialState, State, Statechart, Transition
 
 
 @pytest.fixture
@@ -27,47 +27,3 @@ def state():
     Transition(initial_state, next_state)
     statechart.start()
     return next_state
-
-
-class TestMetadata:
-    def test_create_metadata(self):
-        Metadata()
-
-    def test_activate_inactive_state(self, state):
-        metadata = Metadata()
-        metadata.activate(state.context)
-        metadata.activate(state)
-        assert state in metadata.active_states
-
-    def test_reactivate_active_state(self, state):
-        metadata = Metadata()
-        metadata.activate(state.context)
-        metadata.activate(state)
-        assert state in metadata.active_states
-        metadata.activate(state)
-        assert state in metadata.active_states
-
-    def test_deactivate_active_state(self, state):
-        metadata = Metadata()
-        metadata.activate(state.context)
-        metadata.activate(state)
-        assert state in metadata.active_states
-        metadata.deactivate(state)
-        assert state not in metadata.active_states
-
-    def test_deactivate_inactive_state(self, state):
-        metadata = Metadata()
-        metadata.activate(state.context)
-        metadata.deactivate(state)
-        assert state not in metadata.active_states
-
-    def test_is_active_active_state(self, state):
-        metadata = Metadata()
-        metadata.activate(state.context)
-        metadata.activate(state)
-        assert metadata.is_active(state) is True
-
-    def test_is_active_inactive_state(self, state):
-        metadata = Metadata()
-        metadata.activate(state.context)
-        assert metadata.is_active(state) is False