+++ HISTORY.rst
@@ -26,3 +26,21 @@ History
 ------------------
 
 * Fix internal transition acting like local transition.
+
+0.3.0 (2016-10-16)
+------------------
+
+* Implement display module to generate Plant UML code of a statechart.
+* Raise runtime exception if an action is defined on top level statechart.
+
+0.3.1 (2016-10-16)
+------------------
+
+* Implement specific statechart deactivate function.
+
+0.4.0 (2019-05-18)
+------------------
+
+* Add support for functional action and guard definitions.
+* Deprecate KwEvent, Internal Transitions, Actions and Guard
+* Add support for generating PlantUML diagrams.